Uses payment method labels from API
Retrieves payment method labels from the API and displays them in the dropdown.

This change removes the hardcoded payment method labels and instead fetches them dynamically.  The payment method options are now fetched only once when the component mounts and if methods aren't already loaded.  This also includes a fix for the balance page template key.